backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
hiredis==2.3.2
minio==7.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
httpx==0.26.0
aiohttp==3.9.1
python-docx==1.1.0
PyPDF2==3.0.1
pypdf==4.0.1
reportlab==4.0.7
pydantic==2.5.3
pydantic-settings==2.1.0
